05-story-app/requirements.txt
streamlit
pillow
requests
pandas